+++ src/main/java/team/startup/gwangsan/domain/post/entity/constant/Mode.java
@@ -1,0 +1,6 @@
+package team.startup.gwangsan.domain.post.entity.constant;
+
+public enum Mode {
+    GIVER,
+    RECEIVER
+}
