+++ src/main/java/team/startup/gwangsan/domain/post/entity/constant/ProductStatus.java
@@ -1,0 +1,6 @@
+package team.startup.gwangsan.domain.post.entity.constant;
+
+public enum ProductStatus {
+    ONGOING,
+    COMPLETED
+}
