+++ src/main/java/team/startup/gwangsan/domain/post/entity/constant/Type.java
@@ -1,0 +1,6 @@
+package team.startup.gwangsan.domain.post.entity.constant;
+
+public enum Type {
+    OBJECT,
+    SERVICE
+}
