--- src/main/java/team/startup/gwangsan/post/entity/constant/Mode.java
@@ -1,6 +1,0 @@
-package team.startup.gwangsan.post.entity.constant;
-
-public enum Mode {
-    GIVER,
-    RECEIVER
-}
