--- src/main/java/team/startup/gwangsan/post/entity/constant/ProductStatus.java
@@ -1,6 +1,0 @@
-package team.startup.gwangsan.post.entity.constant;
-
-public enum ProductStatus {
-    ONGOING,
-    COMPLETED
-}
